requirements.txt
appdirs==1.4.4
asgiref==3.3.4
automium==0.2.6
automium-web==0.1.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.4
django-daguerre==3.0.0
django-heroku==0.3.1
django-request==1.5.6
dnspython==2.1.0
email-validator==1.1.2
filelock==3.0.12
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
item==0.0.3
itsdangerous==2.0.1
Jinja2==3.0.1
legacy==0.1.6
market==0.0
MarkupSafe==2.0.1
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
python-dateutil==2.8.1
pytils==0.3
pytz==2021.1
requests==2.25.1
selection==0.0.14
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.15
sqlparse==0.4.1
urllib3==1.26.5
user-agent==0.1.9
virtualenv==20.4.7
weblib==0.1.30
Werkzeug==2.0.1
whitenoise==5.2.0
WTForms==2.3.3
